@@ -4,8 +4,8 @@
   "description": "The plugin for Gulp which allows to pack ArmA pbo files from sources.",
   "main": "dist/main.js",
   "scripts": {
-    "test": "gulp test",
-    "assemble": "gulp assemble"
+    "test": "./node_modules/.bin/gulp test",
+    "assemble": "./node_modules/.bin/gulp assemble"
   },
   "keywords": [
     "arma",
@@ -19,18 +19,21 @@
   "repository": "https://github.com/winseros/gulp-armapbo-plugin",
   "devDependencies": {
     "chai": "^3.5.0",
-    "del": "^2.2.0",
+    "del": "^2.2.2",
     "gulp": "^3.9.1",
-    "gulp-mocha": "^2.2.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-typescript": "^2.13.4",
-    "mocha": "^2.5.1",
-    "sinon": "^1.17.4",
-    "typescript": "^1.8.10",
-    "typings": "^1.0.4"
+    "gulp-mocha": "^3.0.1",
+    "gulp-sourcemaps": "^2.2.0",
+    "gulp-tslint": "^6.1.3",
+    "gulp-typescript": "^3.1.2",
+    "mocha": "^3.1.2",
+    "sinon": "^1.17.6",
+    "tslint": "^3.15.1",
+    "typescript": "^2.0.7",
+    "typings": "^1.5.0",
+    "vinyl": "^2.0.0"
   },
   "dependencies": {
-    "vinyl": "^1.1.1"
+    "vinyl": "^2.0.0"
   },
   "files": [
     "dist/**/*.js",
