@@ -3,17 +3,17 @@ import {IPboHeaderExtension, PboHeaderExtension} from '../domain/pboHeaderExtens
 import {PboWriter} from './pboWriter';
 import {Assert} from '../util/assert';
 import {Sha1} from './sha1';
-import File = require('vinyl');
+import * as File from 'vinyl';
 
 interface EntriesCollection<T> extends Array<T> {
-	size:number;
+	size: number;
 }
 
 export class PboBuilder {
 
 	private static signatureBlockSize = 20;
 
-	build(contents:File[], headerExt:IPboHeaderExtension[]):Buffer {
+	build(contents: File[], headerExt: IPboHeaderExtension[]): Buffer {
 		Assert.isNotNull(contents, 'contents');
 		Assert.isNotNull(headerExt, 'headerExt');
 
@@ -21,8 +21,8 @@ export class PboBuilder {
 		const signature = PboHeaderEntry.getSignatureEntry();
 		size += signature.getSize();
 
-		const headerExtensions= this.getHeaderExtensions(headerExt);
-		size += headerExtensions.size+ 1;//terminating null after the last one
+		const headerExtensions = this.getHeaderExtensions(headerExt);
+		size += headerExtensions.size + 1;//terminating null after the last one
 
 		const headerEntries = this.getHeaderEntries(contents);
 		size += headerEntries.size;
@@ -52,52 +52,52 @@ export class PboBuilder {
 		return result;
 	}
 
-	private getHeaderEntries(contents:File[]):EntriesCollection<IPboHeaderEntry> {
+	private getHeaderEntries(contents: File[]): EntriesCollection<IPboHeaderEntry> {
 		let size = 0;
-		let entries = <EntriesCollection<IPboHeaderEntry>>contents.filter(file => {
+		let entries = contents.filter(file => {
 			return file.contents ? true : false;//filter out directories
 		}).map(file => {
 			const fileSize = file.stat.size;
 			const timeStamp = file.stat.mtime.getTime() / 1000;
 
 			const entry = new PboHeaderEntry(file.relative, PackingMethod.uncompressed, fileSize, timeStamp, fileSize);
-			entry.contents = <Buffer>file.contents;
+			entry.contents = file.contents as Buffer;
 
 			size += entry.getSize() + fileSize;
 
 			return entry;
-		});
+		}) as EntriesCollection<IPboHeaderEntry>;
 
 		entries.size = size;
 		return entries;
 	}
 
-	private getHeaderExtensions(entries:IPboHeaderExtension[]):EntriesCollection<PboHeaderExtension> {
+	private getHeaderExtensions(entries: IPboHeaderExtension[]): EntriesCollection<PboHeaderExtension> {
 		let size = 0;
-		let extensions = <EntriesCollection<PboHeaderExtension>>entries.map(entry => {
+		let extensions = entries.map(entry => {
 			const ext = PboHeaderExtension.fromObject(entry);
 			size += ext.getSize();
 			return ext;
-		});
+		}) as EntriesCollection<PboHeaderExtension>;
 		extensions.size = size;
 		return extensions;
 	}
 
-	private writeHeaderExtensions(writer:PboWriter, extensions:PboHeaderExtension[], buffer:Buffer, offset:number):number {
+	private writeHeaderExtensions(writer: PboWriter, extensions: PboHeaderExtension[], buffer: Buffer, offset: number): number {
 		extensions.forEach(extension => {
 			offset = writer.writeHeaderExtension(buffer, extension, offset);
 		});
 		return offset;
 	}
 
-	private writeHeaderEntries(writer:PboWriter, entries:IPboHeaderEntry[], buffer:Buffer, offset:number):number {
+	private writeHeaderEntries(writer: PboWriter, entries: IPboHeaderEntry[], buffer: Buffer, offset: number): number {
 		entries.forEach(entry => {
 			offset = writer.writeHeaderEntry(buffer, entry, offset);
 		});
 		return offset;
 	}
 
-	private writeContents(entries:IPboHeaderEntry[], buffer:Buffer, offset:number):number {
+	private writeContents(entries: IPboHeaderEntry[], buffer: Buffer, offset: number): number {
 		entries.forEach(entry => {
 			offset += entry.contents.copy(buffer, offset, 0);
 		});
